Update version reference in project
Hi! Thanks a lot for this great project. I noticed the version reference in the readme was not the latest version, but instead of updating it, I made this change as a suggestion. Not sure if you'd rather just update the version number there, but this way you wouldn't have to with each update. Maybe updates don't happen often though :man_shrugging:.